[Challenge Tourial] Fix the compilation error
1. lose precision error caused by 'Env.Students.size()'
2. lib\host\host_function_example\examplefunc.cpp(46,42) : error :'[' tokens introducing attribute are separated by whitespace [-Werror,-Wcompound-token-split-by-space]

+++ include/host/host_function_example/examplefunc.h
@@ -22,7 +22,7 @@ class HostFuncExampleAddStudent
 public:
   HostFuncExampleAddStudent(HostFuncExampleEnvironment &HostEnv)
       : HostFuncExample(HostEnv) {}
-  Expect<uint32_t> body(Runtime::Instance::MemoryInstance *MemInst,
+  Expect<size_t> body(Runtime::Instance::MemoryInstance *MemInst,
                         uint32_t StudentPtr, uint32_t StudentLen);
 };
